feat(quality): complete P6.2 Coverage & Quality checks
Phase 6.2 complete - Coverage 99.5%, all quality checks pass:

✅ P6.2.1: Coverage Report
- Domain + Application: 99.5% (1085 statements, 5 missing)
- Exceeds 90% target by 9.5%

✅ P6.2.2: Coverage Threshold
- All thresholds met
- HTML report generated: htmlcov/index.html

✅ P6.2.3: Linter (Flake8)
- All line length violations fixed
- Clean on 256 source files

✅ P6.2.4: Type Checker (mypy strict)
- Fixed 140 type errors
- Success: no issues found in 256 source files

Type Safety Improvements:
- BarcodeProduct.nutrients now Optional[NutrientProfile] (aligns with GraphQL schema)
- Added null safety checks in barcode_service.py validate_product()
- Fixed calories type: float → int in NutrientProfile
- Added mypy overrides for tests.* modules (disable strict typing)
- Fixed untyped Strawberry decorators with type: ignore[misc]

Quality Configuration:
- pyproject.toml: Added [[tool.mypy.overrides]] for tests
- Tests exempt from disallow_untyped_defs (standard practice)

Cleanup:
- Removed 12 legacy V1 API tests (34 obsolete tests deleted)
- All 605 tests passing

Docs:
- Added REFACTOR/09_PERSISTENCE_STRATEGY.md (MongoDB + factory pattern)
- Updated tracker: Phase 6 now 67% complete (2/3 tasks)

+++ backend/infrastructure/meal/providers/stub_nutrition_provider.py
@@ -85,7 +85,7 @@ class StubNutritionProvider:
         scale_factor = quantity_g / 100.0
 
         return NutrientProfile(
-            calories=base_nutrients["calories"] * scale_factor,
+            calories=int(base_nutrients["calories"] * scale_factor),
             protein=base_nutrients["protein"] * scale_factor,
             carbs=base_nutrients["carbs"] * scale_factor,
             fat=base_nutrients["fat"] * scale_factor,
